Add select city on page profile

+++ src/main/java/com/search_partners/service/CountryAndCityService.java
@@ -9,6 +9,6 @@ public interface CountryAndCityService {
 
     List<Country> getAllCountries();
 
-    List<City> getCities(int id);
+    List<City> getCities(long id);
 
 }
